- broke reusable parts of UI into isolated components
